Separate Solana specific functionality of LedgerUntrustedMessagingTransport into a child class SolanaLedgerUntrustedMessagingTransport
