+++ components/brave_wallet_ui/common/hardware/ledgerjs/ledger-untrusted-transport.test.ts
@@ -20,7 +20,3 @@ test('constructor', async () => {
   )
   expect(untrustedTransport.handlers.size).toEqual(3)
 })
-
-// test('handleUnlock', async () => { })
-// test('handleGetAccount', async () => { })
-// test('handleSignTransaction', async () => { })
